Factor ProjectName to separate file
